Add LNNode vendor switch
